fixing communication between lb and worker
