+++ LoadBalancing/LoadBalancer/QueueClinet.h
@@ -28,6 +28,8 @@ typedef struct {
     QUEUEELEMENT* elements;  // Niz elemenata reda
 } QUEUE;  // Ne zaboravite tačku i zarez na kraju definicije strukture
 
+extern QUEUE* ClientQueue;
+
 // Funkcije za rad sa redom
 QUEUE* init_queue(int capacity);  // Kreira red sa datim kapacitetom
 int is_queue_full(QUEUE* q);      // Proverava da li je red pun
