+++ LoadBalancing/LoadBalancer/x64/Debug/LoadBalancer.log
@@ -1,3 +1,7 @@
-﻿  WorkerCommunication.cpp
-C:\Users\Sasa\Desktop\IKP_PROJEKAT_LOADBALANCER\LoadBalancing\LoadBalancer\WorkerCommunication.cpp(217,23): warning C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data
+﻿  LoadBalancer.cpp
+  ClientCommunication.cpp
+  QueueClient.cpp
+  WorkerCommunication.cpp
+C:\Users\Sasa\Desktop\IKP_PROJEKAT_LOADBALANCER\LoadBalancing\LoadBalancer\WorkerCommunication.cpp(113,23): warning C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data
+  Generating Code...
   LoadBalancer.vcxproj -> C:\Users\Sasa\Desktop\IKP_PROJEKAT_LOADBALANCER\LoadBalancing\x64\Debug\LoadBalancer.exe
